HW1 work on mistakes

+++ Pell.java
@@ -7,7 +7,7 @@ public class Pell {
             if (n < 0 || n > 30) {
                 System.out.println("n должно соответствовать диапазону 0 <= n <= 30");
             } else {
-                int[] array = new int[n + 1];
+                long[] array = new long[n + 1];
                 array[0] = 0;
                 if (n > 0) {
                     array[1] = 1;
